requirements.txt
# Core Flask Application
flask>=3.0.0
werkzeug>=3.0.0

# HTTP Client for API calls
requests>=2.31.0

# HTML Parsing for BBC Scraper
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Production WSGI Server
gunicorn>=21.2.0

# Production Dependencies
python-dotenv>=1.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0

# Monitoring and Logging
structlog>=23.1.0

# Security and Validation
marshmallow>=3.20.0
bleach>=6.0.0

# Database (if needed for future features)
sqlalchemy>=2.0.0

# Development and Testing (only for development)
pytest>=7.4.0
pytest-flask>=1.2.0
black>=23.7.0
flake8>=6.0.0